// generated/src/app/api-models/drg-pricer-revision.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class DrgPricerRevision {
  drgPricerId: string;
  version: string;
  revisionLevel: string;
  description: string;
  effectiveDate: Date;
  termDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}